LE_Kotlin: moved all relevant abstractions of  "geometry" into new dedicated package
