+++ Attt_LocalEngine_Kotlin/src/main/kotlin/gameLogic/GameField.kt
@@ -6,8 +6,8 @@ import constants.SYMBOL_FOR_ABSENT_MARK
 import constants.SYMBOL_FOR_DIVIDER
 import constants.SYMBOL_FOR_NEW_LINE
 import players.PlayerProvider
-import geometry.Coordinates
-import geometry.OneMoveProcessing
+import geometry.abstractions.Coordinates
+import geometry.abstractions.OneMoveProcessing
 import geometry.conceptXY.NearestAreaScanWithXY
 import publicApi.AtttPlayer
 import utilities.Log
