+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/LineDirectionFor1Axis.kt
@@ -1,5 +1,7 @@
 package geometry
 
+import geometry.abstractions.LineDirection
+
 /**
  * this is the simplest case - only for one axis - imagine a line where you can move only in two directions
  */
