+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/abstractions/Coordinates.kt
@@ -1,4 +1,4 @@
-package geometry
+package geometry.abstractions
 
 /**
  * describes any real position which may exist on the game field
