+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/abstractions/GameSpace.kt
@@ -1,4 +1,4 @@
-package geometry
+package geometry.abstractions
 
 /**
  * global describer for any possible space on game field or out its borders - it can be anything about location
