+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/abstractions/OneMoveProcessing.kt
@@ -1,4 +1,4 @@
-package geometry
+package geometry.abstractions
 
 /**
  * abstraction for any action to be done right after every move is made - it analyses the new state of the game field
