+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept2D/NearestAreaScanWith2D.kt
@@ -1,8 +1,8 @@
 package geometry.concept2D
 
-import geometry.Coordinates
+import geometry.abstractions.Coordinates
 import gameLogic.GameField
-import geometry.OneMoveProcessing
+import geometry.abstractions.OneMoveProcessing
 import players.PlayerProvider
 import utilities.Log
