+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept3D/LineDirectionFor3Axes.kt
@@ -1,6 +1,6 @@
 package geometry.concept3D
 
-import geometry.LineDirection
+import geometry.abstractions.LineDirection
 import geometry.LineDirectionFor1Axis
 
 internal data class LineDirectionFor3Axes(
