+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/conceptXY/CoordinatesXY.kt
@@ -1,8 +1,8 @@
 package geometry.conceptXY
 
-import geometry.Border
-import geometry.Coordinates
-import geometry.GameSpace
+import geometry.abstractions.Border
+import geometry.abstractions.Coordinates
+import geometry.abstractions.GameSpace
 
 /**
  * describes every place on the game field, later add 3rd dimension here.
