+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/conceptXY/LineDirectionForXY.kt
@@ -1,6 +1,6 @@
 package geometry.conceptXY
 
-import geometry.LineDirection
+import geometry.abstractions.LineDirection
 
 /**
  * describes all possible directions for the simplest line of 2 dots on a 2d field
